+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Updated default specifications version from 1.3 to 1.4.
 - Changed maximum length of NAACCR IDs from 50 to 32 characters (specifications 1.4).
+
+**Version 5.3 (not released yet)**
+
+- Added new methods on Patient and Tumor to remove a given data item.
 - Updated embedded JRE to Java 12.0.1 (this only affects the standalone program).
 
 **Version 5.2**